表情消息,Emoji工具类优化
